backend2.0/requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
itsdangerous
httpx

google-auth
google-api-python-client
google-generativeai

pydantic

# LangChain specific additions
langchain>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=0.1.0

# For local OAuth flow in scripts
google-auth-oauthlib>=0.5.1 # Or latest
google-auth-httplib2>=0.1.0 # Or latest
